requirements.txt
# ASL Alphabet Detection System - Requirements
# Python 3.8 or higher required

# Computer Vision & Hand Tracking
opencv-python==4.8.1.78
mediapipe==0.10.8

# Machine Learning & Data Processing
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4

# Deep Learning (optional, for neural networks)
# tensorflow==2.15.0  # Uncomment if using TensorFlow models

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Text-to-Speech
pyttsx3==2.90

# Utilities
Pillow==10.1.0

# Development & Testing (optional)
# pytest==7.4.3
# jupyter==1.0.0
# ipykernel==6.27.1
